+++ allensdk/test/brain_observatory/behavior/test_behavior_ophys_session.py
@@ -86,7 +86,7 @@ def test_visbeh_ophys_data_set():
                                  'session_type': 'OPHYS_6_images_B',
                                  'driver_line': ['Camk2a-tTA', 'Slc17a7-IRES2-Cre'],
                                  'behavior_session_uuid': uuid.UUID('69cdbe09-e62b-4b42-aab1-54b5773dfe78'),
-                                 'experiment_datetime': pytz.utc.localize(datetime.datetime(2018, 11, 30, 23, 28, 37)),
+                                 'experiment_datetime': pytz.utc.localize(datetime.datetime(2018, 11, 30, 23, 58, 50, 325000)),
                                  'ophys_frame_rate': 31.0,
                                  'imaging_depth': 375,
                                  'LabTracks_ID': 416369,
